test: s/float/float64/
Otherwise gccgo gives an extra error message not matched by errchk.

R=adg, rsc
CC=golang-dev
https://golang.org/cl/4053042

+++ test/method1.go
@@ -8,10 +8,10 @@ package main
 
 type T struct { }
 func (t *T) M(int, string)	// GCCGO_ERROR "previous"
-func (t *T) M(int, float) { }   // ERROR "redeclared|redefinition"
+func (t *T) M(int, float64) { }   // ERROR "redeclared|redefinition"
 
 func f(int, string)	// GCCGO_ERROR "previous"
-func f(int, float) { }  // ERROR "redeclared|redefinition"
+func f(int, float64) { }  // ERROR "redeclared|redefinition"
 
 func g(a int, b string)  // GCCGO_ERROR "previous"
 func g(a int, c string)  // ERROR "redeclared|redefinition"
